Moved sharedAtom table to Flash.
Made the id a uint32_t so the entries are aligned to 4 bytes as required

+++ m8rscript/Atom.h
@@ -110,7 +110,7 @@ public:
     
     static Atom sharedAtom(SharedAtom id)
     {
-        auto it = _sharedAtomMap.find(static_cast<uint16_t>(id));
+        auto it = _sharedAtomMap.find(static_cast<uint32_t>(id));
         if (it == _sharedAtomMap.end()) {
             return Atom();
         }
